server/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.1
django-admin-ordering==0.11.0
django-appconf==1.0.3
django-ckeditor==5.7.1
django-environ==0.4.5
django-imagekit==4.0.2
django-js-asset==1.2.2
django-storages==1.7.1
djangorestframework==3.9.4
google-api-core==1.12.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gunicorn==19.9.0
honcho==1.0.1
idna==2.8
pilkit==2.0
Pillow==6.0.0
protobuf==3.8.0
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-slugify==3.0.2
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
stringcase==1.2.0
text-unidecode==1.2
urllib3==1.25.3
